ebdjango/requirements.txt
Django==2.1.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
pytz==2020.1
